requirements.txt
aiohappyeyeballs==2.3.5
aiohttp==3.10.2
aiohttp-retry==2.8.3
aiosignal==1.3.1
amqp==5.2.0
ansicon==1.89.0
APScheduler==3.10.4
arabic-reshaper==3.0.0
arrow==1.3.0
asgiref==3.8.1
asn1crypto==1.5.1
async-timeout==4.0.3
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
billiard==4.2.0
blessed==1.20.0
celery==5.4.0
certifi==2024.7.4
cffi==1.16.0
channels==3.0.5
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
constantly==23.10.4
cron-descriptor==1.4.3
cryptography==42.0.5
cssselect2==0.7.0
curlify==2.2.1
daphne==3.0.2
defusedxml==0.7.1
diff-match-patch==20230430
distlib==0.3.8
django==5.1.3
django-allauth==65.0.2
django-celery-beat==2.6.0
django-celery-results==2.5.1
django-cors-headers==4.6.0
django-crispy-forms==2.1
django-crontab==0.7.1
django-environ==0.11.2
django-import-export==3.3.7
django-picklefield==3.2
django-q==1.3.9
django-smart-selects==1.6.0
django-timezone-field==7.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
et-xmlfile==1.1.0
facebook_business==20.0.0
filelock==3.15.4
frozenlist==1.4.1
haversine==2.8.1
html5lib==1.1
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
jinxed==1.2.1
kombu==5.3.7
lxml==5.1.0
MarkupPy==1.14
multidict==6.0.5
mysqlclient==2.2.4
numpy==2.0.1
odfpy==1.4.1
openpyxl==3.1.5
oscrypto==1.3.0
pandas==2.2.2
pillow==10.4.0
platformdirs==4.2.2
prompt_toolkit==3.0.47
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycountry==24.6.1
pycparser==2.21
pyHanko==0.21.0
pyhanko-certvalidator==0.26.3
PyJWT==2.8.0
pyOpenSSL==24.3.0
pyotp==2.9.0
pypdf==4.0.2
pypng==0.20220715.0
python-bidi==0.4.2
python-crontab==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
qrcode==7.4.2
redis==5.0.8
reportlab==4.0.9
requests==2.32.3
service-identity==24.2.0
setuptools==69.2.0
six==1.16.0
sqlparse==0.5.1
stripe==8.4.0
svglib==1.5.1
Tabby==0.4.0
tablib==3.5.0
tinycss2==1.2.1
twilio==9.0.5
Twisted==24.10.0
txaio==23.1.1
types-python-dateutil==2.9.0.20240316
typing_extensions==4.10.0
tzdata==2024.1
tzlocal==5.2
uritools==4.0.2
urllib3==2.2.2
vine==5.1.0
virtualenv==20.26.3
wcwidth==0.2.13
webencodings==0.5.1
xhtml2pdf==0.2.15
xlrd==2.0.1
xlwt==1.3.0
yarl==1.9.4
zope.interface==7.2
